+++ engines/glk/level9/level9_main.h
@@ -87,7 +87,7 @@ struct Bitmap {
 #define L9SETDWORD(x,val) WRITE_LE_UINT32(x, val)
 
 extern byte *startdata;
-extern size_t FileSize;
+extern uint32 FileSize;
 
 extern void level9_initialize();
